requirements.txt
# ASTRAEUS - Autonomous Life Orchestrator
# Python Dependencies

# Telegram Bot
python-telegram-bot[webhooks]==21.0.1
python-telegram-bot[job-queue]==21.0.1

# AI/LLM Clients
openai>=1.12.0
google-generativeai>=0.4.0
ollama>=0.1.7

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.3

# Geographic Calculations
geopy>=2.4.1

# Environment & Configuration
python-dotenv>=1.0.1
pydantic>=2.6.1
pydantic-settings>=2.1.0

# Date/Time Handling
python-dateutil>=2.8.2
pytz>=2024.1

# Async Utilities
asyncio>=3.4.3
aiofiles>=23.2.1

# Logging & Monitoring
structlog>=24.1.0
rich>=13.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0

# Code Quality
ruff>=0.2.1
mypy>=1.8.0

# Whisper for Voice (Optional)
# openai-whisper>=20231117
